+++ src/main/java/kitchenpos/product/application/dto/ProductEventDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.domain;
+package kitchenpos.product.application.dto;
 
 import java.math.BigDecimal;
